api/requirements.txt
fastapi
uvicorn
opentelemetry-sdk
opentelemetry-api
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp-proto-http
